refactor: contants 폴더 구조 재구조화
