+++ src/constants/benefit/badge.js
@@ -1,0 +1,7 @@
+const BADGE = Object.freeze({
+  santa: '산타',
+  tree: '트리',
+  star: '별',
+});
+
+export default BADGE;
