+++ src/constants/benefit/benefitConditions.js
@@ -1,0 +1,5 @@
+const BENEFIT_CONDITIONS = Object.freeze({
+  minAmountForBenefit: 10000,
+});
+
+export default BENEFIT_CONDITIONS;
