--- src/constants/order.js
@@ -1,6 +1,0 @@
-const ORDER = Object.freeze({
-  maxMenuLimit: 20,
-  minAmountForBenefit: 10000,
-});
-
-export default ORDER;
