+++ src/constants/order/orderConditions.js
@@ -1,0 +1,5 @@
+const ORDER_CONDITIONS = Object.freeze({
+  maxMenuLimit: 20,
+});
+
+export default ORDER_CONDITIONS;
