+++ src/constants/order/type.js
@@ -1,0 +1,8 @@
+const TYPE = Object.freeze({
+  appetizer: 'appetizer',
+  main: 'main',
+  dessert: 'dessert',
+  drink: 'drink',
+});
+
+export default TYPE;
